# Symbolic stoichiometric coefficients