--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -38,7 +38,11 @@ end
 # Miscellaneous tests
 #@time @safetestset "Basic Plotting" begin include("plotting.jl") end
 @time @safetestset "Latexify" begin include("latexify.jl") end
-@time @safetestset "Graphs" begin include("graphs.jl") end
+
+# disable on Macs as can't install GraphViz via jll 
+if !Sys.isapple()
+    @time @safetestset "Graphs" begin include("graphs.jl") end
+end
 
 @time @safetestset "Conservation Laws" begin include("conslaws.jl") end
 @time @safetestset "Units" begin include("units.jl") end