--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -9,9 +9,7 @@ using SafeTestsets
 @time @safetestset "ReactionSystem Components Test" begin include("reactionsystem_components.jl") end
 
 # Tests all features realted to constructing a model
-if VERSION >= v"1.6.0"
-    @time @safetestset "1.6 Arrows" begin include("newarrows.jl") end
-end
+@time @safetestset "1.6 Arrows" begin include("newarrows.jl") end
 @time @safetestset "Basic DSL" begin include("dsl.jl") end
 @time @safetestset "Model Construction" begin include("make_model.jl") end
 @time @safetestset "Custom Functions" begin include("custom_functions.jl") end