TODO:
- [x] ODEs and SDEs with tests.
- [x] Jumps with tests ~(needs ModelingToolkit JumpSystem updates)~ edit: Not needed apparently.
- [x] drop requirement for `use_combinatoric_ratelaws=false`.
- [x] benchmark to see if this significantly impacts system construction times.
- [ ] ~DSL support.~ edit: Will save for a future PR as this will be some work...
- [x] tests
- [x] check BCR with jumps for performance regressions on construction.