@isaacsas 2022-03-01
Testing BCR loading with this PR seems to not have any real impact on timings for loading and converting systems:
```julia
Current Catalyst:

prnbng = loadrxnetwork(BNGNetwork(),fname); 
522.819 ms (4224319 allocations: 189.59 MiB)

bosys = convert(ODESystem, rnbng)
2.448 s (26007243 allocations: 1.36 GiB)

boprob = ODEProblem(bosys, Float64[], (0.,tf), Float64[])
1.359 s (13747759 allocations: 662.72 MiB)

—————————————————
PR:
prnbng = loadrxnetwork(BNGNetwork(),fname)
509.176 ms (4224319 allocations: 189.59 MiB)

bosys = convert(ODESystem, rnbng)
2.456 s (26007243 allocations: 1.36 GiB)

boprob = ODEProblem(bosys, Float64[], (0.,tf), Float64[])
1.358 s (13747759 allocations: 662.72 MiB)
```
I forgot to interpolate into the `@btime` calls, but when I reran the PR code interpolating variables it had no real effect.