@isaacsas 2022-03-02
Currently this should work for ODEs and SDEs. I may just update the DSL and then merge this. 

Everything for jumps should be in place here, but we need updates to JumpSystems in ModelingToolkit too which will have to be for another day.