@isaacsas 2022-03-03
BCR JumpSystem construction timings are also unaffected:
```
Jumps
Current Catalyst:
prn = loadrxnetwork(BNGNetwork(), fname)
504.861 ms (4215741 allocations: 189.36 MiB) 

jsys = convert(JumpSystem, rn)
53.115 ms (611173 allocations: 26.10 MiB)

dprob = DiscreteProblem(jsys, u0, (0.0,tequil), [])
6.167 ms (64467 allocations: 2.69 MiB)

jprob = JumpProblem(jsys, dprob, RSSACR(), save_positions=(false,false))
2.703 s (2701719 allocations: 255.82 MiB)

—————————————————
PR:
prn = loadrxnetwork(BNGNetwork(), fname)
501.691 ms (4215741 allocations: 189.36 MiB) 

jsys = convert(JumpSystem, rn)
49.913 ms (611173 allocations: 26.10 MiB)

dprob = DiscreteProblem(jsys, u0, (0.0,tequil), [])
5.995 ms (64467 allocations: 2.69 MiB)

jprob = JumpProblem(jsys, dprob, RSSACR(), save_positions=(false,false))
2.720 s (2701719 allocations: 255.82 MiB)
```