@isaacsas 2022-03-03
@ChrisRackauckas I think this is good to go, but if you could look it over when you have a chance to make sure I'm not doing something bad/dumb in my handling of symbolics that would be great.